fix workflow circular imports
