+++ convex/tsconfig.json
@@ -11,6 +11,7 @@
     "jsx": "react-jsx",
     "skipLibCheck": true,
     "allowSyntheticDefaultImports": true,
+    "noUncheckedIndexedAccess": true,
 
     /* These compiler options are required by Convex */
     "target": "ESNext",
